requirements.txt
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
pefile==2022.5.30
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
python-dotenv==0.21.0
pywin32-ctypes==0.2.0
requests==2.28.1
urllib3==1.26.13
yarl==1.8.2